fix(warg): boot-safe logger entry module; raise sandbox sleepAfter to 5m
- workerd rejects non-handler exports from a worker entry module: move
  BUCKET_KEY_PREFIX/bucketKeyForTs into logger/src/bucket-key.ts so the
  logger boots under wrangler dev/deploy (was a TypeError at startup)
- raise the monolith Sandbox sleepAfter backstop from 30s to 5m: the
  inactivity alarm can fire during an in-flight exec
  (cloudflare/containers#162) and archive jobs run up to ~180s P99, so a
  30s backstop risked killing long jobs mid-run
